refactor: cleanse parameters and field names for customServlet
