+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.connector;
 
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.coyote.request.CustomServlet;
+import org.apache.coyote.servlet.CustomServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,16 +17,16 @@ public class Connector implements Runnable {
     private static final int DEFAULT_PORT = 8080;
     private static final int DEFAULT_ACCEPT_COUNT = 100;
 
-    private final CustomServlet customServlet;
+    private final CustomServlet servlet;
     private final ServerSocket serverSocket;
     private boolean stopped;
 
-    public Connector(CustomServlet customServlet) {
-        this(customServlet, DEFAULT_PORT, DEFAULT_ACCEPT_COUNT);
+    public Connector(CustomServlet servlet) {
+        this(servlet, DEFAULT_PORT, DEFAULT_ACCEPT_COUNT);
     }
 
-    public Connector(final CustomServlet customServlet, final int port, final int acceptCount) {
-        this.customServlet = customServlet;
+    public Connector(final CustomServlet servlet, final int port, final int acceptCount) {
+        this.servlet = servlet;
         this.serverSocket = createServerSocket(port, acceptCount);
         this.stopped = false;
     }
@@ -68,7 +68,7 @@ public class Connector implements Runnable {
             return;
         }
         log.info("connect host: {}, port: {}", connection.getInetAddress(), connection.getPort());
-        var processor = new Http11Processor(connection, customServlet);
+        var processor = new Http11Processor(connection, servlet);
         new Thread(processor).start();
     }
