+++ tomcat/src/main/java/org/apache/catalina/startup/Tomcat.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.startup;
 
 import org.apache.catalina.connector.Connector;
-import org.apache.coyote.request.CustomServlet;
+import org.apache.coyote.servlet.CustomServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,8 +12,7 @@ public class Tomcat {
     private static final Logger log = LoggerFactory.getLogger(Tomcat.class);
 
     public void start() {
-        final var customServlet = new CustomServlet();
-        final var connector = new Connector(customServlet);
+        final var connector = new Connector(new CustomServlet());
         connector.start();
 
         try {
